Implement Trusted Users

+++ mattata.lua
@@ -955,6 +955,8 @@ function mattata.process_spam(message)
             return false, 'The administration plugin is switched off in this chat!'
         elseif not mattata.get_setting(message.chat.id, 'antispam') then
             return false, 'The antispam plugin is switched off in this chat!'
+        elseif mattata.get_setting(message.chat.id, 'trusted permissions antispam') and mattata.is_trusted_user(message.chat.id, message.from.id) then
+            return false, 'That user is a trusted user and he has authorization to spam!'
         end
 
         local msg_count = tonumber(
@@ -1074,43 +1076,47 @@ function mattata:process_message()
         redis:incr('messages:' .. message.from.id .. ':' .. message.chat.id)
     end
     if message.new_chat_members and mattata.get_setting(message.chat.id, 'use administration') and mattata.get_setting(message.chat.id, 'antibot') and not mattata.is_group_admin(message.chat.id, message.from.id) and not mattata.is_global_admin(message.from.id) then
-        local kicked = {}
-        local usernames = {}
-        for k, v in pairs(message.new_chat_members) do
-            if v.username and v.username:lower():match('bot$') and v.id ~= message.from.id and v.id ~= self.info.id and tostring(v.is_bot) == 'true' then
-                local success = mattata.kick_chat_member(message.chat.id, v.id)
-                if success then
-                    table.insert(kicked, mattata.escape_html(v.first_name) .. ' [' .. v.id .. ']')
-                    table.insert(usernames, '@' .. v.username)
-                end
-            end
-        end
-        if #kicked > 0 and #usernames > 0 and #kicked == #usernames then
-            local log_chat = mattata.get_log_chat(message.chat.id)
-            mattata.send_message(log_chat, string.format('<pre>%s [%s] has kicked %s from %s [%s] because anti-bot is enabled.</pre>', mattata.escape_html(self.info.first_name), self.info.id, table.concat(kicked, ', '), mattata.escape_html(message.chat.title), message.chat.id), 'html')
-            return mattata.send_message(message, string.format('Kicked %s because anti-bot is enabled.', table.concat(usernames, ', ')))
-        end
-    end
-    if message.chat.type == 'supergroup' and mattata.get_setting(message.chat.id, 'use administration') and mattata.get_setting(message.chat.id, 'word filter') and not mattata.is_group_admin(message.chat.id, message.from.id) and not mattata.is_global_admin(message.from.id) then
-        local words = redis:smembers('word_filter:' .. message.chat.id)
-        if words and #words > 0 then
-            for k, v in pairs(words) do
-                local text = message.text:lower()
-                if text:match('^' .. v:lower() .. '$') or text:match('^' .. v:lower() .. ' ') or text:match(' ' .. v:lower() .. ' ') or text:match(' ' .. v:lower() .. '$') then
-                    mattata.delete_message(message.chat.id, message.message_id)
-                    local action = mattata.get_setting(message.chat.id, 'ban not kick') and mattata.ban_chat_member or mattata.kick_chat_member
-                    local success = action(message.chat.id, message.from.id)
+        if not mattata.is_trusted_user(message.chat.id, message.from.id) and not mattata.get_setting(message.chat.id, 'trusted permissions antibot') then
+            local kicked = {}
+            local usernames = {}
+            for k, v in pairs(message.new_chat_members) do
+                if v.username and v.username:lower():match('bot$') and v.id ~= message.from.id and v.id ~= self.info.id and tostring(v.is_bot) == 'true' then
+                    local success = mattata.kick_chat_member(message.chat.id, v.id)
                     if success then
-                        if mattata.get_setting(message.chat.id, 'log administrative actions') then
-                            local log_chat = mattata.get_log_chat(message.chat.id)
-                            mattata.send_message(log_chat, string.format('<pre>%s [%s] has kicked %s [%s] from %s [%s] for sending one or more prohibited words.</pre>', mattata.escape_html(self.info.first_name), self.info.id, mattata.escape_html(message.from.first_name), message.from.id, mattata.escape_html(message.chat.title), message.chat.id), 'html')
-                        end
-                        mattata.send_message(message.chat.id, string.format('Kicked %s for sending one or more prohibited words.', message.from.username and '@' .. message.from.username or message.from.first_name))
-                        break_cycle = true
+                        table.insert(kicked, mattata.escape_html(v.first_name) .. ' [' .. v.id .. ']')
+                        table.insert(usernames, '@' .. v.username)
                     end
                 end
             end
-            if break_cycle then return true end
+            if #kicked > 0 and #usernames > 0 and #kicked == #usernames then
+                local log_chat = mattata.get_log_chat(message.chat.id)
+                mattata.send_message(log_chat, string.format('<pre>%s [%s] has kicked %s from %s [%s] because anti-bot is enabled.</pre>', mattata.escape_html(self.info.first_name), self.info.id, table.concat(kicked, ', '), mattata.escape_html(message.chat.title), message.chat.id), 'html')
+                return mattata.send_message(message, string.format('Kicked %s because anti-bot is enabled.', table.concat(usernames, ', ')))
+            end
+        end
+    end
+    if message.chat.type == 'supergroup' and mattata.get_setting(message.chat.id, 'use administration') and mattata.get_setting(message.chat.id, 'word filter') and not mattata.is_group_admin(message.chat.id, message.from.id) and not mattata.is_global_admin(message.from.id) then
+        if not mattata.is_trusted_user(message.chat.id, message.from.id) and not mattata.get_setting(message.chat.id, 'trusted permissions wordfilter') then
+            local words = redis:smembers('word_filter:' .. message.chat.id)
+            if words and #words > 0 then
+                for k, v in pairs(words) do
+                    local text = message.text:lower()
+                    if text:match('^' .. v:lower() .. '$') or text:match('^' .. v:lower() .. ' ') or text:match(' ' .. v:lower() .. ' ') or text:match(' ' .. v:lower() .. '$') then
+                        mattata.delete_message(message.chat.id, message.message_id)
+                        local action = mattata.get_setting(message.chat.id, 'ban not kick') and mattata.ban_chat_member or mattata.kick_chat_member
+                        local success = action(message.chat.id, message.from.id)
+                        if success then
+                            if mattata.get_setting(message.chat.id, 'log administrative actions') then
+                                local log_chat = mattata.get_log_chat(message.chat.id)
+                                mattata.send_message(log_chat, string.format('<pre>%s [%s] has kicked %s [%s] from %s [%s] for sending one or more prohibited words.</pre>', mattata.escape_html(self.info.first_name), self.info.id, mattata.escape_html(message.from.first_name), message.from.id, mattata.escape_html(message.chat.title), message.chat.id), 'html')
+                            end
+                            mattata.send_message(message.chat.id, string.format('Kicked %s for sending one or more prohibited words.', message.from.username and '@' .. message.from.username or message.from.first_name))
+                            break_cycle = true
+                        end
+                    end
+                end
+                if break_cycle then return true end
+            end
         end
     end
     --[[ if message.chat.type == 'supergroup' and mattata.get_setting(message.chat.id, 'use administration') and mattata.get_setting(message.chat.id, 'anti porn') and not mattata.is_group_admin(message.chat.id, message.from.id) and not mattata.is_global_admin(message.from.id) and message.photo then
@@ -1132,15 +1138,17 @@ function mattata:process_message()
         end
     end ]]
     if message.chat.type == 'supergroup' and mattata.get_setting(message.chat.id, 'use administration') and mattata.get_setting(message.chat.id, 'antilink') and not mattata.is_group_admin(message.chat.id, message.from.id) and not mattata.is_global_admin(message.from.id) and mattata.check_links(message) then
-        local action = mattata.get_setting(message.chat.id, 'ban not kick') and mattata.ban_chat_member or mattata.kick_chat_member
-        local success = action(message.chat.id, message.from.id)
-        if success then
-            if mattata.get_setting(message.chat.id, 'log administrative actions') then
-                local log_chat = mattata.get_log_chat(message.chat.id)
-                mattata.send_message(log_chat, string.format('<pre>%s [%s] has kicked %s [%s] from %s [%s] for sending Telegram invite link(s) from unauthorized groups/channels</pre>', mattata.escape_html(self.info.first_name), self.info.id, mattata.escape_html(message.from.first_name), message.from.id, mattata.escape_html(message.chat.title), message.chat.id), 'html')
+        if not mattata.is_trusted_user(message.chat.id, message.from.id) and not mattata.get_setting(message.chat.id, 'trusted permissions antilink') then
+            local action = mattata.get_setting(message.chat.id, 'ban not kick') and mattata.ban_chat_member or mattata.kick_chat_member
+            local success = action(message.chat.id, message.from.id)
+            if success then
+                if mattata.get_setting(message.chat.id, 'log administrative actions') then
+                    local log_chat = mattata.get_log_chat(message.chat.id)
+                    mattata.send_message(log_chat, string.format('<pre>%s [%s] has kicked %s [%s] from %s [%s] for sending Telegram invite link(s) from unauthorized groups/channels</pre>', mattata.escape_html(self.info.first_name), self.info.id, mattata.escape_html(message.from.first_name), message.from.id, mattata.escape_html(message.chat.title), message.chat.id), 'html')
+                end
+                mattata.delete_message(message.chat.id, message.message_id)
+                return mattata.send_message(message.chat.id, string.format('Kicked %s for sending Telegram invite link(s) from unauthorized groups/channels.', message.from.username and '@' .. message.from.username or message.from.first_name))
             end
-            mattata.delete_message(message.chat.id, message.message_id)
-            return mattata.send_message(message.chat.id, string.format('Kicked %s for sending Telegram invite link(s) from unauthorized groups/channels.', message.from.username and '@' .. message.from.username or message.from.first_name))
         end
     end
     if message.new_chat_members and message.chat.type ~= 'private' and mattata.get_setting(message.chat.id, 'use administration') and mattata.get_setting(message.chat.id, 'welcome message') then
